feat: update about me layout and detail
The layout and detail of the AboutContent.vue component has been updated
.The applied changes improved the overall look of the "About Me"
component on the website. Some minor changes were made in the text
layout, and some additional details about the professional background
were introduced for better understanding. This change results in a more
appealing and informative text and visual output.