feat: favicon for dark mode support
Added a stylesheet to the SVG favicon file to support dark mode. This
change adjusts the favicon brightness and stroke color when the user's
system is in a dark color scheme. This will make the icon less harsh on
the eyes of users while using the application in dark mode.