Refactor nuxt.config.ts for concision and clarity
This commit includes several adjustments to enhance the readability and
maintainability of the 'nuxt.config.ts' file. Unnecessary comments and
attributes were removed, such as the redundant 'sizes' attribute from
the 'icon' object and a superfluous comment for the 'srcDir' key. The
'description' and 'keywords' meta tags have also been reworded for
better SEO optimization.