feat: update hyperlink to nuxt-link
Changed the hyperlink tag in 'AboutContent.vue' from a basic HTML anchor
tag to a 'nuxt-link' for better performance and user experience. Nuxt
links are preferred in a Vue.js application because they provide
client-side navigation with the performance benefit of not having to
make a server request upon navigation. This enhances the overall
performance of the website ensuring seamless navigation for users.