feat: add base seo information, og:font
The Nuxt SEO module was added to the list of modules in the Nuxt
configuration file for improving search engine optimization. Moreover,
the metadata of the website was significantly updated. This includes
adding site details and Open Graph image settings for better handling of
 shared links on social media platforms. The web app's page title
 formatting has been updated to use a separator symbol, and a canonical
 URL has been added which can help with SEO by telling search engines
 about the preferred version of the webpage.

Note: ogImage needed ttf file format; woff2 causes errors.