feat: update the Roadmap components
The RoadmapContainer, RoadmapHeader, and RoadmapProjects components have
 been revised to present roadmap data in a semester system instead of a
 quarterly system.

The roadmap data has been altered in `RoadmapContainer.vue` to represent
 two periods, 'h1' (first half) and 'h2' (second half). The changes
 include modifications in the logic related to displaying the current
 roadmap and its corresponding projects. 'RoadmapHeader.vue' is now
 provided with `roadmapTitleOptions` prop, standing for the available
 semesters. The selected title is mapped to its corresponding data in
 'RoadmapContainer'. A new repository 'elite-fitness' is added in `app
 .config.ts`.

The changes aim to render the roadmap presentation more intuitive and
relevant with an academic year structure for viewers especially from
educational and academic past. This is expected to enhance users'
connection and understanding of the flow of events or tasks.