feat: update README for enhanced clarity and added infographics
Enhancements have been made to the README file for better clarity and a
more visually appealing outlook. The introduction and description of
resume access are tweaked for improved readability. Moreover, a
'Website' link has been added that was previously missing. The
'Technologies Used' section is updated with added infographics for a
more comprehensive understanding of the technologies used. Changes in
the icons for the 'Tools & Technologies' section ensure better
relevance.

Lastly, the closing message has been updated to reflect a more
inspiring, future-focused sentiment. These changes aim to enhance the
overall readability and visual appeal of the repository's README.