feat: update footer component styling and layout
Redesigned the footer component with improved layout and interactivity.
The links to GitHub and LinkedIn have been made more prominent and the
visual presentation has been enhanced with a use of gradients and
shadows. The display of site information such as copyright and
technology stack has also been improved for better readability.